Added type 2 data, added workaround for K-39 in fendl-30 that cannot be processed by makxsf on MARCONI.

+++ combine_all.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
-cat xsdir30 > xsdir
-csplit -q xsdir21 '/directory/1'
-cat xx01 >> xsdir
-rm xx??
+# ITER uses fendl21, therefore place these data first
+for T in 1 2; do
+    if [ -f xsdir21.$T ]; then
+        cat xsdir21.$T > xsdir.$T
+    fi
+    if [ -f xsdir30.$T ]; then    
+        csplit -q xsdir30.$T '/directory/1'
+        cat xx01 >> xsdir.$T
+        rm xx??
+    fi        
+done;
